Speed up jarjar with prefix matches
A common use-case for jarjar is to have rule with the format
my.package.name.**, or even my.package.name.MyClass with no wildcards.
Jarjar will convert that pattern to a regular expression and do regular
expression matching on every symbol it scans to determine if the rule
matches. This is very inefficient as most symbols will not match, and a
simple prefix match can rule out the match in most cases.

When parsing rules, extract the plain-text prefix of each rule, which is
everything before the first wildcard. Then index all rules by their
prefix in a prefix trie. This allows:
 - Avoiding slow regular expression matches in most cases, as jarjar
   can immediately observe that there is no matching prefix in the map.
 - Scaling in O(log(n)) with the number of rules when they have distinct
   prefixes (instead of O(n), as a matching prefix can be found from the
   trie instead of looping through all rules.

This was tested to make jarjar faster on the Android codebase (running
all jarjar rules to build a device back-to-back goes from 295s to 270s),
while keeping outputs byte-identical, and providing significant speedups
when many jarjar rules are used.